+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [6.6.0] - 2025-08-26
 ### Added:
 - Updated types to allow for select menus in modals, as well as using a `LabelComponent` as using action rows in modals are now deprecated.
+- Added the `PIN_MESSAGES` permission
 ### Changed:
 - The type `ComponentSelectMenu` is now deprecated in favor of `AnySelectComponent` as each component type is now split out into their own types.
 ## [6.5.0] - 2025-07-17
@@ -628,7 +630,7 @@ This release features mostly completed documentation and changes to the packages
 ## [0.1.0] - 2020-12-15
 - Initial release.
 
-[Unreleased]: https://github.com/Snazzah/slash-create/compare/v6.5.0...HEAD
+[Unreleased]: https://github.com/Snazzah/slash-create/compare/v6.6.0...HEAD
 [0.1.0]: https://github.com/Snazzah/slash-create/releases/tag/v0.1.0
 [0.2.0]: https://github.com/Snazzah/slash-create/compare/v0.1.0...v0.2.0
 [1.0.0]: https://github.com/Snazzah/slash-create/compare/v0.2.0...v1.0.0
@@ -709,3 +711,4 @@ This release features mostly completed documentation and changes to the packages
 [6.4.1]: https://github.com/Snazzah/slash-create/compare/v6.4.0...v6.4.1
 [6.4.2]: https://github.com/Snazzah/slash-create/compare/v6.4.1...v6.4.2
 [6.5.0]: https://github.com/Snazzah/slash-create/compare/v6.4.2...v6.5.0
+[6.6.0]: https://github.com/Snazzah/slash-create/compare/v6.5.0...v6.6.0